Revert "Refactor BenefitImage to separate file"
This reverts commit 27bc6f71022a2893956e7bcd34652c8cb6c6b7d6.
