+++ src/mobility/components/OperatorBenefit.tsx
@@ -2,12 +2,16 @@ import {GenericSectionItem, Section} from '@atb/components/sections';
 import {View, ViewStyle} from 'react-native';
 import {ThemeText} from '@atb/components/text';
 import React from 'react';
+import {StyleSheet} from '@atb/theme';
 import {OperatorBenefitType} from '@atb-as/config-specs/lib/mobility-operators';
 import {getTextForLanguage, useTranslation} from '@atb/translations';
 import {useIsEligibleForBenefit} from '@atb/mobility/use-is-eligible-for-benefit';
 import {FormFactor} from '@atb/api/types/generated/mobility-types_v2';
-import {BenefitImage} from '@atb/mobility/components/BenefitImage';
-import {StyleSheet} from '@atb/theme';
+import {Check} from '@atb/assets/svg/color/icons/status';
+import {
+  BundlingCarSharing,
+  BundlingCityBike,
+} from '@atb/assets/svg/color/images/mobility';
 
 type Props = {
   benefit: OperatorBenefitType;
@@ -45,7 +49,6 @@ export const OperatorBenefit = ({benefit, formFactor, style}: Props) => {
             <BenefitImage
               eligible={isUserEligibleForBenefit}
               formFactor={formFactor}
-              style={styles.benefitImage}
             />
             <View style={styles.benefitContent}>
               {heading && (
@@ -62,6 +65,49 @@ export const OperatorBenefit = ({benefit, formFactor, style}: Props) => {
   );
 };
 
+const BenefitImageAsset = ({formFactor}: {formFactor: FormFactor}) => {
+  const styles = useStyles();
+  switch (formFactor) {
+    case FormFactor.Car:
+      return <BundlingCarSharing style={styles.benefitImage} />;
+    case FormFactor.Bicycle:
+      return <BundlingCityBike style={styles.benefitImage} />;
+    default:
+      return null;
+  }
+};
+
+type BenefitImageProps = {
+  formFactor: FormFactor;
+  eligible: boolean;
+  style?: ViewStyle;
+};
+
+const BenefitImage = ({
+  formFactor,
+  eligible,
+  style,
+}: BenefitImageProps): JSX.Element => {
+  return (
+    <View style={style}>
+      <>
+        <BenefitImageAsset formFactor={formFactor} />
+        {eligible && <BenfitEligibilityIndicator />}
+      </>
+    </View>
+  );
+};
+
+const BenfitEligibilityIndicator = () => {
+  const styles = useStyles();
+
+  return (
+    <View style={styles.indicator}>
+      <Check width={24} height={24} />
+    </View>
+  );
+};
+
 const useStyles = StyleSheet.createThemeHook((theme) => ({
   benefitContainer: {
     flexDirection: 'row',
@@ -73,5 +119,9 @@ const useStyles = StyleSheet.createThemeHook((theme) => ({
   benefitImage: {
     marginEnd: theme.spacings.medium,
     marginStart: theme.spacings.small,
+  },
+  indicator: {
+    position: 'absolute',
+    right: theme.spacings.medium,
   },
 }));
